Trying to fix the CI
